+++ cmd/curator/main.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	version = "0.22.4"
+	version = "0.22.5"
 )
 
 func main() {